feat(decisions): promote Decision detail view into @repowise-dev/ui
Move the decision detail composition (markdown body, evolution lineage,
writable module-link editor, evidence drawer, AI verification prompt,
confirm/undo status actions) into packages/ui behind an injected
DecisionDetailAdapter. Links route through an injected LinkComponent +
href builders, and a renderLinkedIssues slot lets hosted add a Jira
panel (web omits it). Web becomes a thin host; the rendered view is
unchanged. Adds view-level behavioral tests.
