+++ packages/ui/src/decisions/index.ts
@@ -4,3 +4,5 @@ export * from "./verification-badge";
 export * from "./decision-evidence-drawer";
 export * from "./decision-lineage";
 export * from "./decision-graph-view";
+export * from "./decision-detail-adapter";
+export * from "./decision-detail";
